データの取得方法の説明を追加 (#52)
* データの取得方法の説明を追加

* 誤字を修正

* エラーを修正

* データの取得の仕方を修正

* エラーハンドリングを追加

* モックデータをmockディレクトリに移動

* 取得してきたデータの使い方を追加
