+++ package.json
@@ -34,6 +34,7 @@
     "@typescript-eslint/parser": "^5.59.6",
     "autoprefixer": "10.4.14",
     "classcat": "^5.0.4",
+    "encoding": "^0.1.13",
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.4",
     "eslint-config-prettier": "^8.8.0",
