+++ src/app/api/link/route.ts
@@ -5,7 +5,7 @@ import { routeHandlerSupabase } from "@/lib/routeHandlerSupabase";
 export async function GET(request: Request) {
   const supabase = routeHandlerSupabase;
   const { searchParams } = new URL(request.url);
-  const id = searchParams.get("id");
+  const id = await searchParams.get("id");
 
   if (id) {
     const { data } = await supabase.from("link").select("*").eq("id", id);
