+++ src/app/sample/page.tsx
@@ -1,35 +1,32 @@
-import { NextPage } from "next";
+import { mockData } from "@/mock/SampleMockData";
+import { Chef } from "@/types/tableType";
 
 import { ImageCarousel, ImageComponent, ImageGrid } from "@/components/image";
 
-type Mock = {
-  title: string;
-  description?: string;
-  favNum?: number;
-  nameLabel?: string;
-  src: string;
-};
+// fetchする場合はNextPage型は削除してasyncをつける
+const Home = async () => {
+  // サーバーコンポーネントの場合
+  // revalidateは何秒キャッシュされたデータを使うかの設定
+  // 開発中はキャッシュが邪魔になるので0秒に設定
+  const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/chef`, { next: { revalidate: 0 } });
+  if (!res.ok) {
+    throw new Error("データの取得に失敗しました");
+  }
+  const chefs: Chef[] = await res.json();
 
-const Home: NextPage = () => {
-  const mockData: Mock[] = [
-    {
-      title: "トマトとルッコラのマルゲリータトマトとルッコラのマルゲリータトマトとルッコラのマルゲリータ",
-      description: "トマトとルッコラのマルゲリータトマトとルッコラのマルゲリータトマトとルッコラのマルゲリータ",
-      favNum: 2000,
-      nameLabel: "山田シェフ",
-      src: "/images/sample_chef.jpg",
-    },
-    {
-      title: "トマトとルッコラ",
-      description: "トマトとルッコラのマルゲリータ",
-      nameLabel: "武田シェフ",
-      src: "/images/sample_chef.jpg",
-    },
-    ...[...Array(10)].map((_) => ({
-      title: "トマトとルッコラのマルゲリータトマトとルッコラのマルゲリータトマトとルッコラのマルゲリータ",
-      src: "/images/sample_chef.jpg",
-    })),
-  ];
+  // クライアントコンポーネントの場合
+  // const [state, setState] = useState<Chef>();
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  /* { cache: "no-cache" }としないとSSGのような挙動になりデータがキャッシュされる
+   今後、データの追加を試したいときに、キャッシュされているとデータが追加されないのでこのようにすることでキャッシュを無効化して確認ができる */
+  //     const data = await fetch("http://localhost:3000/api/chef", { cache: "no-cache" });
+  //     const chef = await data.json();
+  //     setState(chef);
+  //   };
+  //   fetchData();
+  // }, []);
+  // console.log(state);
 
   return (
     <>
@@ -85,6 +82,18 @@ const Home: NextPage = () => {
         nameLabel="たけちゃんシェフ"
         ratio="3/4"
       />
+      {/* 取得してきたデータの使い方 */}
+      {chefs.map((chef) => (
+        <ImageComponent
+          key={chef.id}
+          src={chef.image_url}
+          alt={`${chef.name}の画像`}
+          nameLabel={chef.name}
+          ratio="3/4"
+          width="large"
+          addClassNames="mb-8"
+        />
+      ))}
     </>
   );
 };
